Move a route helper into its own package
Clips a dependency.
