+++ pkg/cmd/util/route/route.go
@@ -1,4 +1,4 @@
-package util
+package route
 
 import (
 	"fmt"
